requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
munch==2.5.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
python-dateutil
pytz
python-slugify